- multi aggregation implemented - corpus api refactory - bugfix for aggregation in chunk_search
